+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix redefining read-ony `URL.prototype` in NW.js & browsers
 * Fix setting undefined `clearImmediate` function
+* Added `String#repeat(count)` polyfill
 
 ## 0.5.2 (2018-03-02)